simplify: hoist IPC socket path, dedupe fake-primary test helper
